Update Pressure.ino
Script takes a baseline measurement (for example, when on ground), then measures differences in pressure as a difference in height, which is of particular use to us. Eventually we'll need to make a function that allows us to establish 'baseline pressure' during start up sequences, then have another function that relies on this established pressure (since we're running everything off of one large arduino script, then declaring the variable globally should allow other functions to use this established baseline pressure).